fix: Make it possible to pass a folder-path as a command-line arg, instead of requiring a file within that
Changed the command-line path handling logic so that it checks if the folder
is a file or a directory (and if a directory, it will use the first image found
in that folder, if available).

The previous code here would just result in the UI getting stuck, with the
"loading" message being displayed but not resolving.

Known Limitations / TODO's:
* The variable naming is clunky, and could perhaps be tidied up to require
  fewer unpacking steps?
* It would probably be nice to display error messages if the provided paths
  are not suitable to be loaded. Todo comments have been left for those cases.