fix(586): Compare list & scrubber mangles position
Using both the scrubber and the compare list to go to the next or
previous image may clobber the current image position. I refactored the
compare list to store its position as well as removed all clones for the
list.

The result is that the scrubber and compare list maintain their own
positions. I assume that both lists should have their own set of images
(e.g. the scrubber may be open to ~/Pictures while the compare list has
whatever images the user added). This patch does not fix that issue, but
I can work on that in a separate PR if needed.