`last_open_directory` is incorrectly duplicated
Related to #766. It turns out that the `last_open_directory` setting was
saved in two different locations. It's saved in both the user's cache
folder as well as in VolatileSettings. VolatileSettings is used in
several places while the cache folder setting is only used in one place,
so I assume this was an oversight.

As a plus, this simplifies the code in filebrowser.rs a bit too.